+++ spec/active_record/connection_adapters/makara_abstract_adapter_spec.rb
@@ -36,10 +36,10 @@ describe ActiveRecord::ConnectionAdapters::MakaraAbstractAdapter do
     'select release_lock(\'foo\')' => true,
     'select pg_advisory_lock(12345)' => true,
     'select pg_advisory_unlock(12345)' => true
-  }.each do |sql, should_go_to_master|
-    it "determines that \"#{sql}\" #{should_go_to_master ? 'requires' : 'does not require'} master" do
+  }.each do |sql, should_go_to_primary|
+    it "determines that \"#{sql}\" #{should_go_to_primary ? 'requires' : 'does not require'} primary" do
       proxy = klass.new(config(1,1))
-      expect(proxy.master_for?(sql)).to eq(should_go_to_master)
+      expect(proxy.primary_for?(sql)).to eq(should_go_to_primary)
     end
   end
 
@@ -59,7 +59,7 @@ describe ActiveRecord::ConnectionAdapters::MakaraAbstractAdapter do
   }.each do |sql, should_send_to_all_connections|
     it "determines that \"#{sql}\" #{should_send_to_all_connections ? 'should' : 'should not'} be sent to all underlying connections" do
       proxy = klass.new(config(1,1))
-      proxy.master_pool.connections.each{|con| expect(con).to receive(:execute).with(sql).once}
+      proxy.primary_pool.connections.each{|con| expect(con).to receive(:execute).with(sql).once}
       proxy.replica_pool.connections.each do |con|
         if should_send_to_all_connections
           expect(con).to receive(:execute).with(sql).once
@@ -102,7 +102,7 @@ describe ActiveRecord::ConnectionAdapters::MakaraAbstractAdapter do
         max_treats IS NULL
     } => true
   }.each do |sql,should_stick|
-    it "should #{should_stick ? 'stick' : 'not stick'} to master if handling sql like \"#{sql}\"" do
+    it "should #{should_stick ? 'stick' : 'not stick'} to primary if handling sql like \"#{sql}\"" do
       proxy = klass.new(config(0,0))
       expect(proxy.would_stick?(sql)).to eq(should_stick)
     end