+++ spec/active_record/connection_adapters/makara_postgis_adapter_spec.rb
@@ -52,16 +52,16 @@ if RUBY_ENGINE == 'ruby' &&
         end
       end
 
-      it 'should have one master and two replicas' do
-        expect(connection.master_pool.connection_count).to eq(1)
+      it 'should have one primary and two replicas' do
+        expect(connection.primary_pool.connection_count).to eq(1)
         expect(connection.replica_pool.connection_count).to eq(2)
       end
 
       it 'should allow real queries to work' do
         connection.execute('INSERT INTO users (name) VALUES (\'John\')')
 
-        connection.master_pool.connections.each do |master|
-          expect(master).to receive(:execute).never
+        connection.primary_pool.connections.each do |primary|
+          expect(primary).to receive(:execute).never
         end
 
         change_context
@@ -71,7 +71,7 @@ if RUBY_ENGINE == 'ruby' &&
       end
 
       it 'should send SET operations to each connection' do
-        connection.master_pool.connections.each do |con|
+        connection.primary_pool.connections.each do |con|
           expect(con).to receive(:execute).with("SET TimeZone = 'UTC'").once
         end
 
@@ -91,8 +91,8 @@ if RUBY_ENGINE == 'ruby' &&
         connection.execute('SELECT * FROM users')
       end
 
-      it 'should send writes to master' do
-        con = connection.master_pool.connections.first
+      it 'should send writes to primary' do
+        con = connection.primary_pool.connections.first
         expect(con).to receive(:execute).with('UPDATE users SET name = "bob" WHERE id = 1')
         connection.execute('UPDATE users SET name = "bob" WHERE id = 1')
       end
@@ -115,7 +115,7 @@ if RUBY_ENGINE == 'ruby' &&
       end
     end
 
-    context 'with only master connection' do
+    context 'with only primary connection' do
       it 'should not raise errors on read and write' do
         custom_config = config.deep_dup
         custom_config['makara']['connections'].select{|h| h['role'] == 'replica' }.each{|h| h['port'] = '1'}
@@ -135,7 +135,7 @@ if RUBY_ENGINE == 'ruby' &&
         ActiveRecord::Base.clear_all_connections!
 
         custom_config = config.deep_dup
-        custom_config['makara']['connections'].select{|h| h['role'] == 'master' }.each{|h| h['port'] = '1'}
+        custom_config['makara']['connections'].select{|h| h['role'] == 'primary' }.each{|h| h['port'] = '1'}
 
         ActiveRecord::Base.establish_connection(custom_config)