+++ spec/config_parser_spec.rb
@@ -7,8 +7,8 @@ describe Makara::ConfigParser do
       makara: {
         connections: [
           {
-            role: 'master',
-            name: 'themaster'
+            role: 'primary',
+            name: 'theprimary'
           },
           {
             name: 'replica1'
@@ -24,7 +24,7 @@ describe Makara::ConfigParser do
   context '::merge_and_resolve_default_url_config' do
     let(:config_without_url) do
       {
-        master_ttl: 5,
+        primary_ttl: 5,
         blacklist_duration: 30,
         sticky: true,
         adapter: 'mysql2_makara',
@@ -39,7 +39,7 @@ describe Makara::ConfigParser do
 
     let(:config_with_url) do
       {
-        master_ttl: 5,
+        primary_ttl: 5,
         blacklist_duration: 30,
         sticky: true,
         adapter: 'mysql2_makara',
@@ -106,16 +106,16 @@ describe Makara::ConfigParser do
     expect(parser.id).to eq('myproxyidwithreservedcharacters')
   end
 
-  context 'master and replica configs' do
-    it 'should provide master and replica configs' do
+  context 'primary and replica configs' do
+    it 'should provide primary and replica configs' do
       parser = described_class.new(config)
-      expect(parser.master_configs).to eq([
+      expect(parser.primary_configs).to eq([
         {
-          name: 'themaster',
+          name: 'theprimary',
           top_level: 'value',
           sticky: true,
           blacklist_duration: 30,
-          master_ttl: 5
+          primary_ttl: 5
         }
       ])
       expect(parser.replica_configs).to eq([
@@ -124,14 +124,14 @@ describe Makara::ConfigParser do
           top_level: 'value',
           sticky: true,
           blacklist_duration: 30,
-          master_ttl: 5
+          primary_ttl: 5
         },
         {
           name: 'replica2',
           top_level: 'value',
           sticky: true,
           blacklist_duration: 30,
-          master_ttl: 5
+          primary_ttl: 5
         }
       ])
     end
@@ -142,13 +142,13 @@ describe Makara::ConfigParser do
       config[:makara][:connections][1][:top_level] = 'replica value'
 
       parser = described_class.new(config)
-      expect(parser.master_configs).to eq([
+      expect(parser.primary_configs).to eq([
         {
-          name: 'themaster',
+          name: 'theprimary',
           top_level: 'value',
           sticky: true,
           blacklist_duration: 456,
-          master_ttl: 5
+          primary_ttl: 5
         }
       ])
       expect(parser.replica_configs).to eq([
@@ -157,14 +157,14 @@ describe Makara::ConfigParser do
           top_level: 'replica value',
           sticky: true,
           blacklist_duration: 123,
-          master_ttl: 5
+          primary_ttl: 5
         },
         {
           name: 'replica2',
           top_level: 'value',
           sticky: true,
           blacklist_duration: 123,
-          master_ttl: 5
+          primary_ttl: 5
         }
       ])
     end