+++ spec/middleware_spec.rb
@@ -41,15 +41,15 @@ describe Makara::Middleware do
     _, headers, body = middleware.call(env)
 
     expect(headers).to eq({})
-    expect(body).to eq('master/1')
+    expect(body).to eq('primary/1')
   end
 
-  it 'should set the cookie if master is used' do
+  it 'should set the cookie if primary is used' do
     env[:query] = 'update users set name = "phil"'
 
     _, headers, body = middleware.call(env)
 
     expect(headers['Set-Cookie']).to eq("#{key}=mock_mysql%3A#{(now + 5).to_f}; path=/; max-age=10; expires=#{(Time.now + 10).httpdate}; secure; HttpOnly")
-    expect(body).to eq('master/1')
+    expect(body).to eq('primary/1')
   end
 end