+++ spec/pool_spec.rb
@@ -4,7 +4,7 @@ describe Makara::Pool do
   let(:proxy){ FakeProxy.new({makara: pool_config.merge(connections: [])}) }
   let(:pool){ Makara::Pool.new('test', proxy) }
   let(:pool_config){ {blacklist_duration: 5} }
-  let(:master_pool){ Makara::Pool.new('master', proxy) }
+  let(:primary_pool){ Makara::Pool.new('primary', proxy) }
 
   it 'should wrap connections with a ConnectionWrapper as theyre added to the pool' do
     expect(pool.connections).to be_empty
@@ -147,10 +147,10 @@ describe Makara::Pool do
   end
 
   it 'should error out while blacklisted in transaction' do
-    wrapper_a = master_pool.add(pool_config){ FakeConnection.new(open_transactions: 1) }
-    master_pool.add(pool_config){ FakeConnection.new }
+    wrapper_a = primary_pool.add(pool_config){ FakeConnection.new(open_transactions: 1) }
+    primary_pool.add(pool_config){ FakeConnection.new }
     expect {
-      master_pool.provide do |connection|
+      primary_pool.provide do |connection|
         if connection == wrapper_a
           raise Makara::Errors::BlacklistConnection.new(wrapper_a, StandardError.new('failure'))
         end
@@ -158,14 +158,14 @@ describe Makara::Pool do
     }.to raise_error(Makara::Errors::BlacklistedWhileInTransaction)
   end
 
-  it 'skips blacklisted connections in master pool when not in transaction' do
-    wrapper_a = master_pool.add(pool_config){ FakeConnection.new(open_transactions: 0) }
-    master_pool.add(pool_config){ FakeConnection.new }
-    master_pool.provide do |connection|
+  it 'skips blacklisted connections in primary pool when not in transaction' do
+    wrapper_a = primary_pool.add(pool_config){ FakeConnection.new(open_transactions: 0) }
+    primary_pool.add(pool_config){ FakeConnection.new }
+    primary_pool.provide do |connection|
       if connection == wrapper_a
         raise Makara::Errors::BlacklistConnection.new(wrapper_a, StandardError.new('failure'))
       end
     end
-    10.times{ master_pool.provide{|connection| expect(connection).not_to eq(wrapper_a) } }
+    10.times{ primary_pool.provide{|connection| expect(connection).not_to eq(wrapper_a) } }
   end
 end