+++ spec/proxy_spec.rb
@@ -3,82 +3,82 @@ require 'spec_helper'
 describe Makara::Proxy do
   let(:klass){ FakeProxy }
 
-  it 'sets up a master and replica pool no matter the number of connections' do
+  it 'sets up a primary and replica pool no matter the number of connections' do
     proxy = klass.new(config(0, 0))
-    expect(proxy.master_pool).to be_a(Makara::Pool)
+    expect(proxy.primary_pool).to be_a(Makara::Pool)
     expect(proxy.replica_pool).to be_a(Makara::Pool)
 
     proxy = klass.new(config(2, 0))
-    expect(proxy.master_pool).to be_a(Makara::Pool)
+    expect(proxy.primary_pool).to be_a(Makara::Pool)
     expect(proxy.replica_pool).to be_a(Makara::Pool)
 
     proxy = klass.new(config(0, 2))
-    expect(proxy.master_pool).to be_a(Makara::Pool)
+    expect(proxy.primary_pool).to be_a(Makara::Pool)
     expect(proxy.replica_pool).to be_a(Makara::Pool)
 
     proxy = klass.new(config(2, 2))
-    expect(proxy.master_pool).to be_a(Makara::Pool)
+    expect(proxy.primary_pool).to be_a(Makara::Pool)
     expect(proxy.replica_pool).to be_a(Makara::Pool)
   end
 
   it 'instantiates N connections within each pool' do
     proxy = klass.new(config(1, 2))
 
-    expect(proxy.master_pool.connection_count).to eq(1)
+    expect(proxy.primary_pool.connection_count).to eq(1)
     expect(proxy.replica_pool.connection_count).to eq(2)
   end
 
-  it 'should delegate any unknown method to a connection in the master pool' do
+  it 'should delegate any unknown method to a connection in the primary pool' do
     proxy = klass.new(config(1, 2))
 
-    con = proxy.master_pool.connections.first
+    con = proxy.primary_pool.connections.first
     allow(con).to receive(:irespondtothis){ 'hello!' }
 
     expect(proxy).to respond_to(:irespondtothis)
     expect(proxy.irespondtothis).to eq('hello!')
   end
 
-  describe '#stick_to_master' do
+  describe '#stick_to_primary' do
     let(:proxy) { klass.new(config(1, 2)) }
 
-    it 'should use master if manually forced' do
-      expect(proxy.master_for?('select * from users')).to eq(false)
+    it 'should use primary if manually forced' do
+      expect(proxy.primary_for?('select * from users')).to eq(false)
 
-      proxy.stick_to_master!
+      proxy.stick_to_primary!
 
-      expect(proxy.master_for?('select * from users')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(true)
     end
 
     it 'should persist stickiness by default' do
       now = Time.now
-      proxy.stick_to_master!
+      proxy.stick_to_primary!
 
       next_context = Makara::Context.next
       expect(next_context[proxy.id]).to be >= (now + 5).to_f
 
       proxy = klass.new(config(1, 2))
-      expect(proxy.master_for?('select * from users')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(true)
     end
 
     it 'optionally skips stickiness persistence, so it applies only to the current request' do
       now = Time.now
-      proxy.stick_to_master!(false)
+      proxy.stick_to_primary!(false)
 
-      expect(proxy.master_for?('select * from users')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(true)
       next_context = Makara::Context.next
       expect(next_context).to be_nil # Nothing to persist, so context is empty
 
       proxy = klass.new(config(1, 2))
-      expect(proxy.master_for?('select * from users')).to eq(false)
+      expect(proxy.primary_for?('select * from users')).to eq(false)
     end
 
-    it 'supports a float master_ttl for stickiness duration' do
+    it 'supports a float primary_ttl for stickiness duration' do
       now = Time.now
       config = config(1, 2).dup
-      config[:makara][:master_ttl] = 0.5
+      config[:makara][:primary_ttl] = 0.5
       proxy = klass.new(config)
 
-      proxy.stick_to_master!
+      proxy.stick_to_primary!
 
       next_context = Makara::Context.next
       expect(next_context[proxy.id]).to be >= (now + 0.5).to_f
@@ -94,80 +94,80 @@ describe Makara::Proxy do
     end
 
     it 'should provide the replica pool for a read' do
-      expect(proxy.master_for?('select * from users')).to eq(false)
+      expect(proxy.primary_for?('select * from users')).to eq(false)
     end
 
-    it 'should provide the master pool for a write' do
-      expect(proxy.master_for?('insert into users values (a,b,c)')).to eq(true)
+    it 'should provide the primary pool for a write' do
+      expect(proxy.primary_for?('insert into users values (a,b,c)')).to eq(true)
     end
 
-    # master is used, it should continue being used for the duration of the context
-    it 'should stick to master once used for a sticky operation' do
-      expect(proxy.master_for?('insert into users values (a,b,c)')).to eq(true)
-      expect(proxy.master_for?('select * from users')).to eq(true)
+    # primary is used, it should continue being used for the duration of the context
+    it 'should stick to primary once used for a sticky operation' do
+      expect(proxy.primary_for?('insert into users values (a,b,c)')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(true)
     end
 
-    it 'should not stick to master if stickiness is disabled' do
+    it 'should not stick to primary if stickiness is disabled' do
       proxy.sticky = false
-      expect(proxy.master_for?('insert into users values (a,b,c)')).to eq(true)
-      expect(proxy.master_for?('select * from users')).to eq(false)
+      expect(proxy.primary_for?('insert into users values (a,b,c)')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(false)
     end
 
-    it 'should not stick to master if we are in a without_sticking block' do
-      expect(proxy.master_for?('insert into users values (a,b,c)')).to eq(true)
+    it 'should not stick to primary if we are in a without_sticking block' do
+      expect(proxy.primary_for?('insert into users values (a,b,c)')).to eq(true)
 
       proxy.without_sticking do
-        expect(proxy.master_for?('insert into users values (a,b,c)')).to eq(true)
-        expect(proxy.master_for?('select * from users')).to eq(false)
+        expect(proxy.primary_for?('insert into users values (a,b,c)')).to eq(true)
+        expect(proxy.primary_for?('select * from users')).to eq(false)
       end
 
-      expect(proxy.master_for?('select * from users')).to eq(true)
-      expect(proxy.master_for?('insert into users values (a,b,c)')).to eq(true)
-      expect(proxy.master_for?('select * from users')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(true)
+      expect(proxy.primary_for?('insert into users values (a,b,c)')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(true)
     end
 
-    it 'should not stick to master after without_sticking block if there is a write in it' do
-      expect(proxy.master_for?('select * from users')).to eq(false)
+    it 'should not stick to primary after without_sticking block if there is a write in it' do
+      expect(proxy.primary_for?('select * from users')).to eq(false)
 
       proxy.without_sticking do
-        expect(proxy.master_for?('insert into users values (a,b,c)')).to eq(true)
-        expect(proxy.master_for?('select * from users')).to eq(false)
+        expect(proxy.primary_for?('insert into users values (a,b,c)')).to eq(true)
+        expect(proxy.primary_for?('select * from users')).to eq(false)
       end
 
-      expect(proxy.master_for?('select * from users')).to eq(false)
+      expect(proxy.primary_for?('select * from users')).to eq(false)
     end
 
-    it "should not release master if it was stuck in the same request (no context changes yet)" do
-      expect(proxy.master_for?('insert into users values (a,b,c)')).to eq(true)
-      expect(proxy.master_for?('select * from users')).to eq(true)
+    it "should not release primary if it was stuck in the same request (no context changes yet)" do
+      expect(proxy.primary_for?('insert into users values (a,b,c)')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(true)
 
       Timecop.travel Time.now + 10 do
-        # master_ttl has passed but we are still in the same request, so current context
+        # primary_ttl has passed but we are still in the same request, so current context
         # is still relevant
-        expect(proxy.master_for?('select * from users')).to eq(true)
+        expect(proxy.primary_for?('select * from users')).to eq(true)
       end
     end
 
-    it 'should release master if all stuck connections are released' do
-      expect(proxy.master_for?('insert into users values (a,b,c)')).to eq(true)
-      expect(proxy.master_for?('select * from users')).to eq(true)
+    it 'should release primary if all stuck connections are released' do
+      expect(proxy.primary_for?('insert into users values (a,b,c)')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(true)
 
       Makara::Context.release_all
 
-      expect(proxy.master_for?('select * from users')).to eq(false)
+      expect(proxy.primary_for?('select * from users')).to eq(false)
     end
 
-    it 'should use master if all replicas are blacklisted' do
+    it 'should use primary if all replicas are blacklisted' do
       allow(proxy.replica_pool).to receive(:completely_blacklisted?){ true }
-      expect(proxy.master_for?('select * from users')).to eq(true)
+      expect(proxy.primary_for?('select * from users')).to eq(true)
     end
 
-    it 'should use master if all replicas become blacklisted as part of the invocation' do
+    it 'should use primary if all replicas become blacklisted as part of the invocation' do
       allow(proxy.replica_pool).to receive(:next).and_return(nil)
 
       test = double
       expect(test).to receive(:blacklisting).once
-      expect(test).to receive(:using_master).once
+      expect(test).to receive(:using_primary).once
 
       proxy.send(:appropriate_pool, :execute, ['select * from users']) do |pool|
         if pool == proxy.replica_pool
@@ -176,7 +176,7 @@ describe Makara::Proxy do
           pool.connections.each(&:_makara_blacklist!)
           pool.provide
         else
-          test.using_master
+          test.using_primary
         end
       end
     end
@@ -187,21 +187,21 @@ describe Makara::Proxy do
       # weird setup to allow for the correct
       proxy.replica_pool.connections.each(&:_makara_blacklist!)
       proxy.replica_pool.instance_variable_get('@blacklist_errors') << StandardError.new('some replica connection issue')
-      proxy.master_pool.connections.each(&:_makara_blacklist!)
-      proxy.master_pool.instance_variable_get('@blacklist_errors') << StandardError.new('some master connection issue')
+      proxy.primary_pool.connections.each(&:_makara_blacklist!)
+      proxy.primary_pool.instance_variable_get('@blacklist_errors') << StandardError.new('some primary connection issue')
 
-      allow(proxy).to receive(:_appropriate_pool).and_return(proxy.replica_pool, proxy.master_pool)
+      allow(proxy).to receive(:_appropriate_pool).and_return(proxy.replica_pool, proxy.primary_pool)
 
       begin
         proxy.send(:appropriate_pool, :execute, ['select * from users']) do |pool|
           pool.provide{|c| c }
         end
       rescue Makara::Errors::AllConnectionsBlacklisted => e
-        expect(e.message).to eq('[Makara/master] All connections are blacklisted -> some master connection issue -> [Makara/replica] All connections are blacklisted -> some replica connection issue')
+        expect(e.message).to eq('[Makara/primary] All connections are blacklisted -> some primary connection issue -> [Makara/replica] All connections are blacklisted -> some replica connection issue')
       end
 
       proxy.replica_pool.connections.each{|con| expect(con._makara_blacklisted?).to eq(false) }
-      proxy.master_pool.connections.each{|con| expect(con._makara_blacklisted?).to eq(false) }
+      proxy.primary_pool.connections.each{|con| expect(con._makara_blacklisted?).to eq(false) }
     end
   end
 end